+++ shared/rx-allocate-core.js
@@ -498,6 +498,7 @@
     addTeamColumn: Lab.addTeamColumn,
     stageMove: Lab.stageMove,
     stageMoves: Lab.stageMoves,
+    unstageIds: Lab.unstageIds,
     draftSummary: Lab.draftSummary,
     homeColumnKey: Lab.homeColumnKey,
     placementReason: Lab.placementReason,